% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{add_project_subset}
\alias{add_project_subset}
\title{Add a Subset to a REDCap Database}
\usage{
add_project_subset(
  project,
  subset_name,
  filter_field,
  filter_choices,
  dir_other = file.path(project$dir_path, "output"),
  file_name = paste0(project$short_name, "_", subset_name),
  form_names = NULL,
  field_names = NULL,
  deidentify = TRUE,
  reset = FALSE
)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{subset_name}{Character. The name of the subset to create.}

\item{filter_field}{Character. The name of the field in the database to filter on.}

\item{filter_choices}{Vector. The values of \code{filter_field} used to define the subset.}

\item{dir_other}{Character. The directory where the subset file will be saved.
Default is the \code{output} folder within the database directory.}

\item{file_name}{Character. The base name of the file where the subset will be saved.
Default is \verb{<project$short_name>_<subset_name>}.}

\item{form_names}{Character vector. Names of forms to include in the subset. Default is \code{NULL}, which includes all forms.}

\item{field_names}{Character vector. Names of specific fields to include in the subset. Default is \code{NULL}, which includes all fields.}

\item{deidentify}{Logical. Whether to deidentify the data in the subset. Default is \code{TRUE}.}

\item{reset}{Logical. If \code{TRUE}, overwrite existing subset files with the same name. Default is \code{FALSE}.}
}
\value{
A modified \code{project} object that includes the newly created subset.
The subset is also saved as a file in the specified directory.
}
\description{
Creates a subset of the main REDCap database (\code{project}) based on specific filter criteria
and saves it to a specified directory. The subset can be further customized with
additional forms, fields, and deidentification options.
}
\details{
This function filters the main REDCap database using the specified \code{filter_field}
and \code{filter_choices}, then creates a new subset with optional deidentification.
It can be customized to include only specific forms or fields. The resulting subset
is saved to a file for future use.
}
\seealso{
\code{\link{save_project}} for saving the main database or subsets.
}
