% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_files.R
\name{add_REDCap_folder}
\alias{add_REDCap_folder}
\title{Uploads a folder name to REDCap}
\usage{
add_REDCap_folder(project, name)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{name}{folder name}
}
\value{
messages for confirmation
}
\description{
Uploads a folder name to REDCap
}
