% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_remove_project}
\alias{cache_remove_project}
\title{Remove project from cache}
\usage{
cache_remove_project(short_name)
}
\value{
data.frame of projects from the cache
}
\description{
This will remove a project from cache. Remember cache only stores information
like short_name, token_name, directory location, and more from setup_project.
If you want to truly delete the project files go the folder you setup.
}
\seealso{
Other Project Cache Functions: 
\code{\link{cache_clear}()},
\code{\link{cache_path}()},
\code{\link{get_projects}()}
}
\concept{Project Cache Functions}
\keyword{Cache}
\keyword{Functions}
\keyword{Project}
