% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{check_folder_for_projects}
\alias{check_folder_for_projects}
\title{List File Paths of REDCapDB Projects in a Folder}
\usage{
check_folder_for_projects(file_path, validate = T)
}
\arguments{
\item{file_path}{Character. The path to the folder to search.}

\item{validate}{Logical. If \code{TRUE}, the function will only accept valid directories previously set up with \code{setup_DB()}. Default is \code{TRUE}.}
}
\value{
A character vector of file paths for valid REDCapDB project files in the folder. Returns an empty character vector if no valid files are found.
}
\description{
Searches a specified folder for files related to REDCapDB projects and returns their file paths.
Optionally validates the folder to ensure it was previously set up using \code{setup_DB()}.
}
\details{
This function checks a folder (and optionally validates its setup) for \code{.RData} files that correspond to REDCapDB projects.
It identifies files with the extension \code{.RData} and names ending in \verb{_RosyREDCap}, filtering out any unrelated files.
}
\seealso{
\link{setup_DB} for setting up valid directories.
}
