% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{clear_project_fields}
\alias{clear_project_fields}
\title{clear_project_fields}
\usage{
clear_project_fields(project)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Clears and project fields that were added with \code{add_project_fields()}
or if \code{setup_project()} contained \code{add_default_fields = TRUE}
}
