% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{clear_project_transformation}
\alias{clear_project_transformation}
\title{clear_project_transformation
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
clear_project_transformation(project)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\description{
Clears and project fields that were added with \code{add_project_fields()} and
form_transformations added with \code{add_project_transformation()} ...
or if \code{setup_project()} contained \code{add_default_fields = TRUE} or
\code{add_default_fields = TRUE}
}
