% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{.config}
\alias{.config}
\title{Configuration}
\format{
An object of class \code{list} of length 7.
}
\usage{
.config # internal object
}
\description{
Internal configuration helpers used to retrieve package configuration
values from options or environment variables.

Configuration is resolved in the following order:
\enumerate{
\item \code{getOption("redcapsync.config.option.name")}
\item \code{Sys.getenv("REDCAPSYNC_CONFIG_OPTION_NAME")} # skipped for functions
\item Default if unable to find and validate from above.
}
}
\details{
\subsection{allow.test.names}{

Logical for \code{\link[=setup_project]{setup_project()}} allowing \code{project_name} starting with TEST_.
Default is \code{FALSE}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set with options (which will be prioritized over envvar)
getOption("redcapsync.config.allow.test.names") # get
options(redcapsync.config.allow.test.names = FALSE) # set

# set with ennvar (which will be prioritized when options not defined)
Sys.getenv("REDCAPSYNC_CONFIG_ALLOW_TEST_NAMES")  # get
Sys.setenv(REDCAPSYNC_CONFIG_ALLOW_TEST_NAMES = FALSE) # or set in .Renviron
}\if{html}{\out{</div>}}
}

\subsection{show.api.messages}{

Logical for showing display API messages from REDCapR and redcapAPI.
Default is \code{FALSE}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set with options (which will be prioritized over envvar)
getOption("redcapsync.config.show.api.messages") # get
options(redcapsync.config.show.api.messages = FALSE) # set

# set with ennvar (which will be prioritized when options not defined)
Sys.getenv("REDCAPSYNC_CONFIG_SHOW_API_MESSAGES") # get
Sys.setenv(REDCAPSYNC_CONFIG_SHOW_API_MESSAGES = FALSE) # or set in .Renviron
}\if{html}{\out{</div>}}
}

\subsection{verbose}{

Logical for showing display API messages from REDCapR and redcapAPI.
Default is \code{FALSE}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set with options (which will be prioritized over envvar)
getOption("redcapsync.config.verbose") # get
options(redcapsync.config.verbose = FALSE) # set

# set with ennvar (which will be prioritized when options not defined)
Sys.getenv("REDCAPSYNC_CONFIG_VERBOSE") # get
Sys.setenv(REDCAPSYNC_CONFIG_VERBOSE = FALSE) # or set in .Renviron
}\if{html}{\out{</div>}}
}

\subsection{offline}{

Logical for offline, which if TRUE will block any API calls.
Default is \code{FALSE}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set with options (which will be prioritized over envvar)
getOption("redcapsync.config.offline") # get
options(redcapsync.config.offline = FALSE) # set

# set with ennvar (which will be prioritized when options not defined)
Sys.getenv("REDCAPSYNC_CONFIG_OFFLINE") # get
Sys.setenv(REDCAPSYNC_CONFIG_OFFLINE = FALSE) # or set in .Renviron
}\if{html}{\out{</div>}}
}

\subsection{cache.dir}{

Character file path overriding the default cache directory. Default follow
system standards via rappdir, hoardr, or R_USER_CACHE_DIR

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set with options (which will be prioritized over envvar)
getOption("redcapsync.config.cache.dir") # get
options(redcapsync.config.cache.dir = "file/path/to/keep/cache") # set

# set with ennvar (which will be prioritized when options not defined)
Sys.getenv("REDCAPSYNC_CONFIG_CACHE_DIR") # get
Sys.setenv(REDCAPSYNC_CONFIG_CACHE_DIR = "file/path/to/keep/cache") # set
}\if{html}{\out{</div>}}
}

\subsection{header.style}{

Excel sheet header using \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set with options (which will be prioritized over envvar)
getOption("redcapsync.config.header.style") # get
options(
 redcapsync.config.header.style = openxlsx::createStyle(fontSize = 12L)
)
# does not look up envvar due to it's function type
}\if{html}{\out{</div>}}
}

\subsection{body.style}{

Excel sheet body using \code{\link[openxlsx:createStyle]{openxlsx::createStyle()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set with options (which will be prioritized over envvar)
getOption("redcapsync.config.body.style") # get
options(redcapsync.config.body.style = openxlsx::createStyle(fontSize = 12L))
# does not look up envvar due to it's function type
}\if{html}{\out{</div>}}
}
}
\section{Option Names}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{option_list <- list(
  redcapsync.config.allow.test.names = NULL,
  redcapsync.config.show.api.messages = NULL,
  redcapsync.config.verbose = NULL,
  redcapsync.config.offline = NULL,
  redcapsync.config.cache.dir = NULL,
  redcapsync.config.header.style = NULL,
  redcapsync.config.body.style = NULL
)
}\if{html}{\out{</div>}}
}

\section{Environment Variable Names}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{envvar_list <- list(
  REDCAPSYNC_CONFIG_ALLOW_TEST_NAMES = NA,
  REDCAPSYNC_CONFIG_SHOW_API_MESSAGES = NA,
  REDCAPSYNC_CONFIG_VERBOSE = NA,
  REDCAPSYNC_CONFIG_OFFLINE = NA,
  R_USER_CACHE_DIR = NA, # affects your entire cache for any package
  REDCAPSYNC_CONFIG_CACHE_DIR = NA # affects only REDCapSync Cache
)
}\if{html}{\out{</div>}}
}

\seealso{
See \code{vignette("Tokens", package = "REDCapSync")}
}
\keyword{internal}
