% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_helpers.R
\name{deidentify_project}
\alias{deidentify_project}
\title{Deidentify the REDCap Database}
\usage{
deidentify_project(
  project,
  identifiers = NULL,
  drop_free_text = FALSE,
  date_handling = "none"
)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{identifiers}{Optional character vector of column names that should be
excluded from the \code{project}. If not provided, fields where
\code{project$metadata$fields$identifier == "y"} will be used as the default.}

\item{drop_free_text}{Logical. If \code{TRUE}, columns containing free text
will also be excluded from the \code{project}. Default is \code{FALSE}.}

\item{date_handling}{character string. One of \code{none},\code{lowest-overall-zero},
\code{lowest-record-zero}, \code{shuffle-record-randomly}, or zero date date in form of
\code{2012-12-05}}
}
\value{
A \code{project} object with deidentified forms.
}
\description{
Removes identifying information from the REDCap object (\code{project}).
This can be done either based on the \code{identifier} field in the metadata or
by specifying custom identifiers.
}
\details{
This function modifies the \code{project} object to exclude specified identifiers
or any columns flagged as identifiers in the metadata. Free-text fields can
also be optionally removed, ensuring the resulting dataset complies with
deidentification standards.
}
\seealso{
\code{\link{save_project}} for saving the modified database.
}
