% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_files.R
\name{delete_REDCap_file}
\alias{delete_REDCap_file}
\title{Uploads a folder name to REDCap}
\usage{
delete_REDCap_file(project, doc_id)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{doc_id}{from the file list \code{get_REDCap_file_repository(project)}}
}
\value{
messages for confirmation
}
\description{
Uploads a folder name to REDCap
}
