% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{filter_project}
\alias{filter_project}
\title{Select REDCap Records from project}
\usage{
filter_project(
  project,
  transform,
  filter_field,
  filter_choices,
  form_names,
  field_names,
  warn_only = FALSE,
  no_duplicate_cols = FALSE
)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{transform}{A logical flag (\code{TRUE} or \code{FALSE}). If \code{TRUE}, filter using transformation. Defaults to \code{FALSE}.}

\item{filter_field}{A character string representing an extra variable name to be filtered by. This field must be present in the data frame.}

\item{filter_choices}{A character vector of values to filter by for the \code{filter_field}. Only records with these values in the specified field will be included.}

\item{form_names}{A character vector of form names to be included in the filtered data. If missing, all forms are included.}

\item{field_names}{A character vector of field names to be included in the filtered data. If missing, all fields are included.}

\item{warn_only}{A logical flag (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function will issue a warning instead of stopping if the filtering criteria do not match any records. Defaults to \code{FALSE}.}

\item{no_duplicate_cols}{A logical flag (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function will avoid including duplicate columns in the output. Defaults to \code{FALSE}.}
}
\value{
A modified \code{project} object with filtered records and columns based on the provided criteria.
}
\description{
This function filters the records in the \code{project} object by specified
criteria, such as field names, form names, and optional filtering based on a
specific field and its values. It returns a modified \code{project} object
containing only the records that match the filter criteria.
}
\details{
This function filters the data in the \code{project} object according to the specified form and field names and optional filter criteria. If no field names or form names are provided, it defaults to using all fields and forms in the database.
The function uses the helper \code{filter_DF_list} to apply the filtering logic to the \code{project$data} list.
}
