% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{generate_project_summary_test}
\alias{generate_project_summary_test}
\title{Select REDCap Records from project}
\usage{
generate_project_summary_test(
  project,
  transform,
  filter_field,
  filter_choices,
  filter_list = NULL,
  filter_strict = TRUE,
  field_names = NULL,
  form_names = NULL,
  no_duplicate_cols = FALSE,
  deidentify = TRUE,
  drop_free_text = FALSE,
  date_handling = "none",
  upload_compatible = TRUE,
  clean = TRUE,
  drop_blanks = TRUE,
  include_metadata = TRUE,
  annotate_metadata = TRUE,
  include_record_summary = TRUE,
  include_users = TRUE,
  include_log = TRUE
)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{transform}{Logical. Whether to transform the data in the subset.
Default
is \code{TRUE}.}

\item{filter_field}{Character. The name of the field in the database to
filter on.}

\item{filter_choices}{Vector. The values of \code{filter_field} used to define the
subset.}

\item{filter_list}{Vector. The values of \code{filter_field} used to define the
subset.}

\item{filter_strict}{Logical. If \code{TRUE}, all forms will be filtered by
criteria. If \code{FALSE}, will convert original filter to id column and filter
all other forms by that record. Default is \code{TRUE}.}

\item{field_names}{Character vector. Names of specific fields to include in
the subset. Default is \code{NULL}, which includes all fields.}

\item{form_names}{Character vector. Names of forms to include in the subset.
Default is \code{NULL}, which includes all forms.}

\item{no_duplicate_cols}{A logical flag (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the
function will avoid including duplicate columns in the output. Defaults to
\code{FALSE}.}

\item{deidentify}{Logical. Whether to deidentify the data in the subset.
Default is \code{TRUE}.}

\item{drop_free_text}{Logical. If \code{TRUE}, columns containing free text
will also be excluded from the \code{project}. Default is \code{FALSE}.}

\item{date_handling}{character string. One of \code{none},\code{lowest-overall-zero},
\code{lowest-record-zero}, \code{shuffle-record-randomly}, or zero date date in form of
\code{2012-12-05}}

\item{upload_compatible}{Logical. If \code{TRUE}, the data will be compatible with
REDCap API upload. The main conflict is numeric or date variables in a
project with missing codes while \code{clean} = \code{TRUE}. R converts these to \code{NA}.
Default is \code{TRUE}.}

\item{clean}{Logical. If \code{TRUE}, the data will be cleaned before summarizing.
Default is \code{TRUE}.}

\item{drop_blanks}{Logical. If \code{TRUE}, records with blank fields will be
dropped. Default is \code{TRUE}.}

\item{include_metadata}{Logical. If \code{TRUE}, metadata will be included in the
summary. Default is \code{TRUE}.}

\item{annotate_metadata}{Logical. If \code{TRUE}, metadata will be annotated in
the summary. Default is \code{TRUE}.}

\item{include_record_summary}{Logical. If \code{TRUE}, a record summary will be
included in the generated summary. Default is \code{TRUE}.}

\item{include_users}{Logical. If \code{TRUE}, user-related information will be
included in the summary. Default is \code{TRUE}.}

\item{include_log}{Logical. If \code{TRUE}, the log of changes will be included in
the summary. Default is \code{TRUE}.}
}
\value{
A modified \code{project} object with filtered records and columns based
on the provided criteria.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function filters the records in the \code{project} object by specified
criteria, such as field names, form names, and optional filtering based on a
specific field and its values. It returns a modified \code{project} object
containing only the records that match the filter criteria.
}
\details{
This function filters the data in the \code{project} object according to the
specified form and field names and optional filter criteria. If no field
names or form names are provided, it defaults to using all fields and forms
in the database. The function uses the helper \code{filter_form_list} to apply the
filtering logic to the \code{project$data} list.
}
