% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{generate_summary_by_name}
\alias{generate_summary_by_name}
\title{Generate a Summary from a Subset Name}
\usage{
generate_summary_by_name(project, subset_name)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{subset_name}{Character. The name of the subset from which to generate
the summary.}
}
\value{
A list containing the generated summary based on the specified options. The
list includes filtered and cleaned data, metadata, and other summary details.
}
\description{
Generates a summary from a predefined subset of data within a REDCap project.
The summary can be customized based on various options, such as cleaning the
data, including metadata, and annotating metadata.
}
\details{
This function allows you to generate a summary of data from a specific subset
of records within the REDCap project. The function provides flexible options
for cleaning, annotating, and including metadata, as well as controlling
whether to include record summaries, user information, and logs.
}
