% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{get_dir}
\alias{get_dir}
\title{get your directory}
\usage{
get_dir(project)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\value{
file path of directory
}
\description{
get your directory
}
