% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_projects}
\alias{get_projects}
\title{Get your REDCap projects used by REDCapSync}
\usage{
get_projects()
}
\value{
data.frame of projects from the cache

data.frame of cached projects
}
\description{
Everytime a project is synced, basic project information is saved to the
package user cache so that \link{load_project} and \link{sync} work across R sessions.
}
\details{
No direct project data is stored in the cache. Notably, tokens and data are
not stored here. The key variables stored in the cache are...
\itemize{
\item \code{project_name} - unique identifier for REDCapSync package
\item \code{redcap_uri} - server location
\item \code{token_name} - where to find token environment with \code{\link[=Sys.getenv]{Sys.getenv()}}
\item \code{dir_path} - where to saved project and associated files locally
\item \code{project_id} - obtained from API call and "locks-in" the connection
\item \code{redcap_version} - obtained from API call and affects links
\item \code{last_sync} and \code{sync_frequency} - informs REDCap sync of when to update
\item other variables from project info and some internal package mechanics
}
}
\seealso{
\code{vignette("Cache", package = "REDCapSync")}

Other Cache Functions: 
\code{\link{cache_clear}()}
}
\concept{Cache Functions}
\keyword{Cache}
