% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapR.R
\name{get_REDCap_report}
\alias{get_REDCap_report}
\title{Get REDCap Report}
\usage{
get_REDCap_report(DB, report_id, silent = TRUE)
}
\arguments{
\item{DB}{A validated \code{DB} object containing REDCap project data and
settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{report_id}{character or integer of REDCap report ID. This can be found at the end of the URL of the report.}
}
\value{
data.frame of REDCap report
}
\description{
Get REDCap Report
}
