% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_summarize.R
\name{labelled_to_raw_form}
\alias{labelled_to_raw_form}
\title{Clean to Raw REDCap forms}
\usage{
labelled_to_raw_form(FORM, project)
}
\arguments{
\item{FORM}{data.frame of labelled REDCap to be converted to raw REDCap (for uploads)}

\item{project}{A validated \code{project} object containing REDCap project data and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\value{
project object that has been filtered to only include the specified records
}
\description{
Clean to Raw REDCap forms
}
