% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_helpers.R
\name{get_project_url}
\alias{get_project_url}
\alias{get_record_url}
\title{Open Links to REDCap Pages}
\usage{
get_project_url(project, link_type = "home", open_browser = TRUE)

get_record_url(project, record, page, instance, open_browser = TRUE)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and settings. Generated using \code{project <- \link{load_project}("PROJ")}
or \link{setup_project}()}

\item{link_type}{choose one of "base", "home", "record_home",
"records_dashboard", "api", "api_playground", "codebook", "user_rights",
"setup", "logging", "designer", "dictionary", "data_quality", "identifiers"}

\item{open_browser}{logical for launching the link in internet browser}

\item{record}{REDCap record id or study id etc, any column names that match
\code{project$metadata$id_col}}

\item{page}{REDCap page for the record. Must be one of
\code{project$metadata$forms$form_name}}

\item{instance}{REDCap instance if it's a repeating instrument}
}
\value{
internet link
}
\description{
Opens browser page for a given project object.
}
\details{
Uses \code{\link[utils:browseURL]{utils::browseURL()}} to open the specified REDCap page.
In order for the function to work you must have ran
\code{project <- sync_project(project)} successfully at least once.
If the link brings you to a page that doesn't work check the URL. It's
possible your institution may have changed redcap versions, which is part of
the URL. In that case run \code{project <- sync_project(project)} again.
You may have to be signed into REDCap for it to work properly.
When in doubt, just seek out the page by navigating on your own in REDCap.
Report issues if you can.
}
\keyword{internal}
