% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{nav_to_dir}
\alias{nav_to_dir}
\title{nav_to_dir}
\usage{
nav_to_dir(project)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\value{
opens browser link
}
\description{
nav_to_dir
}
