% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{raw_to_labelled_form}
\alias{raw_to_labelled_form}
\title{Raw to Labelled REDCap forms}
\usage{
raw_to_labelled_form(form, project)
}
\arguments{
\item{form}{data.frame of raw REDCap to be converted to labelled REDCap}

\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\value{
project object
}
\description{
Raw to Labelled REDCap forms
}
