% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapSync-package.R
\docType{package}
\name{REDCapSync-package}
\alias{REDCapSync}
\alias{REDCapSync-package}
\title{REDCapSync: Encapsulated REDCap Project Objects for Syncronized Data Pipelines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Wraps dozens of REDCap API endpoints into a standardized R6 object. Research Electronic Data Capture (REDCap) is a survey and database web application software maintained by Vanderbilt University. It has a robust application programming Interface (API) utilized by several R packages. 'REDCapSync' uses 'redcapAPI' and 'REDCapR' behind-the-scenes to retrieve all metadata, data, and log details for a project. To minimize unneccesary server calls, the interim REDCap log is used to retrieve only recently updated records. The user can define subsets, derived fields, and transformations that save to a user-defined directory. Having a standarized, API-effecient, project-agnostic R object streamlines downstream use in scripts, functions, and shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thecodingdocs/REDCapSync}
  \item \url{https://thecodingdocs.github.io/REDCapSync/}
  \item Report bugs at \url{https://github.com/thecodingdocs/REDCapSync/issues}
}

}
\author{
\strong{Maintainer}: Brandon Rose \email{thecodingdocs@gmail.com} (\href{https://orcid.org/0009-0009-7813-1960}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Natalie Goulett [contributor]
  \item Chris Hilsinger-Pate [contributor]
  \item Nikola Susic [contributor]
  \item Mason Thornton [contributor]
  \item Sydney Stern (\href{https://orcid.org/0000-0002-1479-605X}{ORCID}) [contributor]
}

}
\keyword{internal}
