% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-REDCapSync_project.R
\name{REDCapSync_project}
\alias{REDCapSync_project}
\title{Setup or Load REDCapSync Project}
\value{
An R6ClassGenerator
}
\description{
\link[R6:R6Class]{R6} project object for \link{REDCapSync}
Main class for managing REDCap data, metadata, and sync operations.
Users should construct objects using \code{\link[=setup_project]{setup_project()}}.
}
\seealso{
\link{setup_project} for initializing the \code{project} object.'
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-REDCapSync_project-new}{\code{REDCapSync_project$new()}}
\item \href{#method-REDCapSync_project-info}{\code{REDCapSync_project$info()}}
\item \href{#method-REDCapSync_project-add_summary}{\code{REDCapSync_project$add_summary()}}
\item \href{#method-REDCapSync_project-generate_summary}{\code{REDCapSync_project$generate_summary()}}
\item \href{#method-REDCapSync_project-add_field}{\code{REDCapSync_project$add_field()}}
\item \href{#method-REDCapSync_project-sync}{\code{REDCapSync_project$sync()}}
\item \href{#method-REDCapSync_project-save}{\code{REDCapSync_project$save()}}
\item \href{#method-REDCapSync_project-show_metadata}{\code{REDCapSync_project$show_metadata()}}
\item \href{#method-REDCapSync_project-show_data}{\code{REDCapSync_project$show_data()}}
\item \href{#method-REDCapSync_project-use}{\code{REDCapSync_project$use()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-new"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-new}{}}}
\subsection{Method \code{new()}}{
The end user will not see \code{project$new()}. This is handled internally.
Users should construct objects using \code{\link[=setup_project]{setup_project()}}. The remain methods
will be accessible to any user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$new(project)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{a list object meant to be stored internally within R6}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-info"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-info}{}}}
\subsection{Method \code{info()}}{
Print project metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-add_summary"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-add_summary}{}}}
\subsection{Method \code{add_summary()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$add_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-generate_summary"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-generate_summary}{}}}
\subsection{Method \code{generate_summary()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$generate_summary(summary_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_name}}{Character of summary_name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-add_field"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-add_field}{}}}
\subsection{Method \code{add_field()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$add_field()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-sync"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-sync}{}}}
\subsection{Method \code{sync()}}{
Updates the REDCap database (\code{project} object) by fetching the latest data
from the REDCap server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$sync(
  summarize = TRUE,
  save_to_dir = TRUE,
  hard_check = FALSE,
  hard_reset = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summarize}}{Logical (TRUE/FALSE). If TRUE, summarizes data to directory.}

\item{\code{save_to_dir}}{Logical (TRUE/FALSE). If TRUE, saves the updated data to
the directory. Default is \code{TRUE}.}

\item{\code{hard_check}}{Will check REDCap even if not due (see \code{sync_frequency}
parameter from \code{setup_project()})}

\item{\code{hard_reset}}{Logical that forces a fresh update if TRUE. Default is
\code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function updates the REDCap database by fetching the latest data from
the REDCap server. It supports various options such as forcing a fresh
update, checking logs for a specified number of days, and retrieving files
from REDCap. The function can also handle metadata-only updates and batch
processing.
}

\subsection{Returns}{
Messages for confirmation.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-save"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-save}{}}}
\subsection{Method \code{save()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-show_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-show_metadata}{}}}
\subsection{Method \code{show_metadata()}}{
Returns list of data or the specified form.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$show_metadata(type = "fields")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{string of either "fields","forms", or "choices"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-show_data"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-show_data}{}}}
\subsection{Method \code{show_data()}}{
Returns list of data or the specified form.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$show_data(form)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form}}{string of raw form name such as "survey_one"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSync_project-use"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSync_project-use}{}}}
\subsection{Method \code{use()}}{
returns internal list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSync_project$use()}\if{html}{\out{</div>}}
}

}
}
