% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-REDCapSyncProject.R
\name{REDCapSyncProject}
\alias{REDCapSyncProject}
\title{Setup or Load REDCapSync Project}
\value{
A \code{REDCapSyncProject} R6 object.
}
\description{
Project class (REDCapSync)

Main class for managing REDCap data, metadata, and sync operations.

Users should construct objects using \code{\link[=setup_project]{setup_project()}}, not using
\code{REDCapSyncProject$new()} directly.
}
\section{Private Methods}{


\describe{
\item{\code{load(short_name)}}{Load project internals.}
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-REDCapSyncProject-new}{\code{REDCapSyncProject$new()}}
\item \href{#method-REDCapSyncProject-info}{\code{REDCapSyncProject$info()}}
\item \href{#method-REDCapSyncProject-add_summary}{\code{REDCapSyncProject$add_summary()}}
\item \href{#method-REDCapSyncProject-generate_summary}{\code{REDCapSyncProject$generate_summary()}}
\item \href{#method-REDCapSyncProject-add_field}{\code{REDCapSyncProject$add_field()}}
\item \href{#method-REDCapSyncProject-sync}{\code{REDCapSyncProject$sync()}}
\item \href{#method-REDCapSyncProject-save}{\code{REDCapSyncProject$save()}}
\item \href{#method-REDCapSyncProject-use}{\code{REDCapSyncProject$use()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncProject-new"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncProject-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this REDCapSyncProject  \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncProject$new(short_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{short_name}}{Character. The project short identifier.}

\item{\code{description}}{short_name Character project identifier.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncProject-info"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncProject-info}{}}}
\subsection{Method \code{info()}}{
Print project metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncProject$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncProject-add_summary"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncProject-add_summary}{}}}
\subsection{Method \code{add_summary()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncProject$add_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncProject-generate_summary"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncProject-generate_summary}{}}}
\subsection{Method \code{generate_summary()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncProject$generate_summary(short_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{short_name}}{Character. The project short identifier.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncProject-add_field"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncProject-add_field}{}}}
\subsection{Method \code{add_field()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncProject$add_field()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncProject-sync"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncProject-sync}{}}}
\subsection{Method \code{sync()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncProject$sync()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncProject-save"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncProject-save}{}}}
\subsection{Method \code{save()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncProject$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncProject-use"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncProject-use}{}}}
\subsection{Method \code{use()}}{
returns internal list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncProject$use()}\if{html}{\out{</div>}}
}

}
}
