% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{rmarkdown_project}
\alias{rmarkdown_project}
\title{rmarkdown_project}
\usage{
rmarkdown_project(project, dir_other)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{dir_other}{Character string specifying the directory where the
RMarkdown report will be saved. If not provided, it defaults to the \code{output}
directory inside the project's main directory.}
}
\value{
A message indicating the creation of the RMarkdown report and the
path to the generated file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Generates an RMarkdown report for the given REDCap database (\code{project}
object). This function creates an RMarkdown file in the specified directory
or default directory, allowing users to create custom reports based on the
database content.
}
\details{
This function checks if a directory is specified, and if not, defaults to the
\code{output} folder within the project's directory. It generates the RMarkdown
file that can then be used for further processing or rendering into HTML,
PDF, or other formats.
}
\seealso{
\code{\link[REDCapSync]{save_project}} for saving the \code{project} object.

Other db_functions: 
\code{\link{sync_all}()},
\code{\link{sync_project}()}
}
\concept{db_functions}
