% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{run_quality_checks}
\alias{run_quality_checks}
\title{Run Quality Checks}
\usage{
run_quality_checks(project)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\value{
project object
}
\description{
Run Quality Checks
}
