% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{save_project}
\alias{save_project}
\title{Save or Delete project file from the directory}
\usage{
save_project(project, silent = FALSE)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\value{
Message
}
\description{
This will save/delete the "<short_name>_REDCapSync.RData" file in the given
project
directories R_objects folder. These are optional functions given that
\code{save_project} is a also handled by a default parameter in \code{sync_project.}
}
\details{
delete_project will not delete any other files from that directory.
The
user must delete any other files manually.
}
\seealso{
Other project object: 
\code{\link{setup_project}()}
}
\concept{project object}
\keyword{internal}
