% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{summarize_project}
\alias{summarize_project}
\title{Summarize REDCap Database}
\usage{
summarize_project(project, reset = FALSE)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{reset}{Logical (TRUE/FALSE). If TRUE, forces the summary generation
even if there are issues. Default is \code{FALSE}.}
}
\value{
List. Returns a list containing the summarized data, including
records, metadata, users, logs, and any other specified data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Summarizes the REDCap database (\code{project} object) by filtering and generating
a summary list.
}
\details{
This function filters the REDCap database based on the provided parameters
and generates a summary list. The summary can include metadata, record
summaries, user information, and logs. The function also supports
deidentification and cleaning of the data.
}
\seealso{
\link{setup_project} for initializing the \code{project} object.
\link{sync_project} for updating the \code{project} object.

Other db_functions: 
\code{\link{rmarkdown_project}()},
\code{\link{sync_project}()},
\code{\link{transform_project}()}
}
\concept{db_functions}
