% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_project.R
\name{sync_multiple}
\alias{sync_multiple}
\title{Synchronize REDCap Data}
\usage{
sync_multiple(
  short_names = NULL,
  summarize = TRUE,
  hard_check = FALSE,
  hard_reset = FALSE,
  silent = FALSE
)
}
\arguments{
\item{short_names}{character vector of project short_names previously setup.
If = NULL, will get all from \code{get_projects()}}
}
\value{
invisible
}
\description{
Updates the REDCap database (\code{project} object) by fetching the latest data
from the REDCap server.
}
\details{
syncs multiple projects as defined. Will not load projects
}
\seealso{
\link{setup_project} for initializing the \code{project} object.
}
