% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_tokens.R
\name{test_REDCap_token}
\alias{test_REDCap_token}
\title{Test REDCap API Token linked to a DB Object}
\usage{
test_REDCap_token(DB, set_if_fails = T, launch_browser = T)
}
\arguments{
\item{DB}{A validated \code{DB} object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{set_if_fails}{Logical (TRUE/FALSE). If TRUE and test connection fails, asks user to paster token into consult using \code{set_REDCap_token(DB)} function. Default is \code{TRUE}.}

\item{launch_browser}{Logical (TRUE/FALSE). If TRUE, launches the REDCap login page in the default web browser when validation fails. Default is \code{TRUE}.}
}
\value{
Logical. Returns \code{TRUE} if the API token is valid, otherwise \code{FALSE}.
}
\description{
Validates the REDCap API token stored in the \code{DB} object by attempting a connection to the REDCap server.
}
\details{
This function tests whether the API token stored in the \code{DB} object is valid by making a request to the REDCap server.
If the token is invalid, the function can optionally open the REDCap login page in a browser (\code{launch_browser}) and/or reset the token (\code{set_if_fails}) using the console.
}
\seealso{
\href{../articles/Tokens.html}{pkgdown article on tokens}
\href{https://thecodingdocs.github.io/REDCapDB/articles/Tokens.html}{pkgdown article on tokens}

Other Token Functions: 
\code{\link{set_REDCap_token}()},
\code{\link{view_REDCap_token}()}
}
\concept{Token Functions}
\keyword{Functions}
\keyword{Token}
