% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_project.R
\name{untransform_project}
\alias{untransform_project}
\title{Untransform project}
\usage{
untransform_project(project, allow_partial = FALSE)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{allow_partial}{Logical. If TRUE, allows filtered projects with less than original field names. Default is FALSE.}
}
\value{
The original, untransformed project object.
}
\description{
This function reverses any transformations applied to a REDCap database object.
If the database is not transformed, it will return the original project without changes.
The function can optionally allow partial transformations if specified.
}
