% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_update.R
\name{update_DB}
\alias{update_DB}
\title{Update REDCap Database}
\usage{
update_DB(
  DB,
  set_token_if_fails = TRUE,
  force = FALSE,
  day_of_log = 10,
  labelled = TRUE,
  get_files = FALSE,
  original_file_names = FALSE,
  entire_log = FALSE,
  metadata_only = FALSE,
  ask_about_overwrites = TRUE,
  save_to_dir = TRUE,
  records = NULL,
  batch_size = 2000
)
}
\arguments{
\item{DB}{A validated \code{DB} object containing REDCap project data and
settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{set_token_if_fails}{Logical (TRUE/FALSE). If TRUE, prompts the user to
set the REDCap API token if the update fails. Default is \code{TRUE}.}

\item{force}{Logical that forces a fresh update if TRUE. Default is \code{FALSE}.}

\item{day_of_log}{Integer. Number of days to be checked in the log. Default
is \code{10}.}

\item{labelled}{Logical (TRUE/FALSE). If TRUE, returns labelled REDCap data. If FALSE, returns raw data. Default is \code{TRUE}.}

\item{get_files}{Logical (TRUE/FALSE). If TRUE, retrieves files from REDCap. Default is \code{FALSE}.}

\item{original_file_names}{Logical (TRUE/FALSE). If TRUE, uses original file names for retrieved files. Default is \code{FALSE}.}

\item{entire_log}{Logical (TRUE/FALSE). If TRUE, retrieves the entire log. Default is \code{FALSE}.}

\item{metadata_only}{Logical (TRUE/FALSE). If TRUE, updates only the metadata. Default is \code{FALSE}.}

\item{ask_about_overwrites}{Logical (TRUE/FALSE). If TRUE, prompts the user before overwriting existing data. Default is \code{TRUE}.}

\item{save_to_dir}{Logical (TRUE/FALSE). If TRUE, saves the updated data to the directory. Default is \code{TRUE}.}

\item{records}{optional records character vector}

\item{batch_size}{Integer. Number of records to process in each batch. Default is \code{2000}.}
}
\value{
Messages for confirmation.
}
\description{
Updates the REDCap database (\code{DB} object) by fetching the latest data from
the REDCap server.
}
\details{
This function updates the REDCap database by fetching the latest data from
the REDCap server. It supports various options such as forcing a fresh
update, checking logs for a specified number of days, and retrieving files
from REDCap. The function can also handle metadata-only updates and batch
processing.
}
\seealso{
\link{setup_DB} for initializing the \code{DB} object.

Other db_functions: 
\code{\link{rmarkdown_DB}()},
\code{\link{summarize_DB}()},
\code{\link{transform_DB}()}
}
\concept{db_functions}
