% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_upload.R
\name{upload_form_to_REDCap}
\alias{upload_form_to_REDCap}
\title{Upload to REDCap}
\usage{
upload_form_to_REDCap(to_be_uploaded, project, batch_size = 500)
}
\arguments{
\item{to_be_uploaded}{data.frame in raw coded form. If you worked with clean
data pass your data to \code{labelled_to_raw_form(form,project)} first.}

\item{project}{A validated \code{project} object containing REDCap project data and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{batch_size}{numeric of how big the REDCap batch upload is. Default 500.}
}
\value{
messages
}
\description{
This will only overwrite and new data. It will not directly delete and data.
Because this is the only function that can mess up your data, use it at your
own risk.
Remember all changes are saved in the redcap log if there's an issue. Missing
rows and columns are fine!
}
