% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_upload.R
\name{upload_project_to_REDCap}
\alias{upload_project_to_REDCap}
\title{Upload from your directory to REDCap}
\usage{
upload_project_to_REDCap(
  project,
  batch_size = 500,
  ask = TRUE,
  view_old = TRUE,
  n_row_view = 20
)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{batch_size}{Numeric. The number of records to upload in each batch.
Default is 500.}

\item{ask}{Logical. If TRUE, the function will prompt you to preview the data
that will be uploaded before proceeding. Defaults to TRUE.}

\item{view_old}{Logical. If TRUE, the function will show a preview of the old
records before upload. Defaults to TRUE.}

\item{n_row_view}{Numeric. The number of rows to display when previewing old
data. Default is 20.}
}
\value{
A series of messages indicating the progress and status of the
upload.
}
\description{
This function is designed to upload changes from a locally modified \code{project}
object to REDCap. It should be run after using
\code{project_import <- read_from_REDCap_upload(project)}.
The function compares the imported data (\code{project_import}) to the existing
data in \code{project} and only uploads new or changed data. It will not directly
delete any data.

This function has the potential to modify your data, so it should be used
cautiously. Any changes made through this function will be logged in the
REDCap log.
}
\details{
This function uploads changes to a REDCap project, based on the differences
between the locally imported data (\code{project_import}) and the existing data
(\code{project}).
It uploads changes in batches as specified by \code{batch_size} and allows you to
preview the changes before the upload if \code{ask} is set to TRUE.
It will not delete any data from REDCap, and it is intended to only upload
new or modified records. This function should be used with caution. Any
changes during the upload will be logged in the REDCap system log.
}
