% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_project.R
\name{upload_transform_to_project}
\alias{upload_transform_to_project}
\title{upload_transform_to_project Transform}
\usage{
upload_transform_to_project(project)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\value{
project object
}
\description{
upload_transform_to_project Transform
}
