% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_clear}
\alias{cache_clear}
\title{Clear your cached projects}
\usage{
cache_clear(project_names = NULL)
}
\arguments{
\item{project_names}{character vector of project project_names previously
setup. If NULL, will get all from \code{get_projects()}}
}
\value{
message of outcome
}
\description{
Finds the location of the cache established by \code{\link[hoardr]{hoard}}
and deletes stored project information (not data)! If you provide
\code{project_names}, it will remove only those projects from the cache.
}
\details{
The cache only stores information like project_name, token_name, directory
location, and other details from \code{\link[=setup_project]{setup_project()}}. If you want to truly delete
the project files, you must do so at the project directory you set up.
}
\examples{
\dontrun{
cache_clear("OLD_PROJECT")
cache_clear() # every project
}
}
\seealso{
\code{vignette("Cache", package = "REDCapSync")}
}
\concept{Cache Functions}
\keyword{Cache}
\keyword{Functions}
