% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{clear_project_summaries}
\alias{clear_project_summaries}
\title{clear_project_summaries}
\usage{
clear_project_summaries(project, summary_names = NULL)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}
}
\value{
invisible return of project and a message
}
\description{
clears project summaries
}
\keyword{internal}
