% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapSyncDataset.R
\name{dataset}
\alias{dataset}
\alias{REDCapSyncDataset}
\title{REDCapSync Encapsulated Datsaset Object}
\value{
An R6ClassGenerator which is used internally to create or load a
dataset object for the user
}
\description{
\link[R6:R6Class]{R6} project object for \link{REDCapSync}
}
\seealso{
\link{setup_project} for initializing the \code{project} object.'
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list of data where names are forms}

\item{\code{metadata}}{list of metadata}

\item{\code{records}}{data.frame of records with timestamps}

\item{\code{users}}{data.frame of users with timestamps}

\item{\code{log}}{data.frame of log}

\item{\code{comments}}{data.frame of comments}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{project_name}}{Read-only character string of project_name as
assigned with \link{setup_project}.}

\item{\code{dataset_name}}{Read-only character string of dataset_name}

\item{\code{n_records}}{Read-only integer of n_records}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-REDCapSyncDataset-new}{\code{REDCapSyncDataset$new()}}
\item \href{#method-REDCapSyncDataset-print}{\code{REDCapSyncDataset$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncDataset-new}{}}}
\subsection{Method \code{new()}}{
The end user will not see \code{dataset$new()}. This is handled internally.
Users should construct objects using \link{REDCapSyncProject}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncDataset$new(project, dataset_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{a list object meant to be stored internally within R6}

\item{\code{dataset_name}}{dataset name setup with \code{project$add_dataset}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-REDCapSyncDataset-print"></a>}}
\if{latex}{\out{\hypertarget{method-REDCapSyncDataset-print}{}}}
\subsection{Method \code{print()}}{
Print some key dataset information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{REDCapSyncDataset$print()}\if{html}{\out{</div>}}
}

}
}
