% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_project.R
\name{generate_project_summary}
\alias{generate_project_summary}
\title{Generate a Summary from a Subset Name}
\usage{
generate_project_summary(
  project,
  summary_name,
  transformation_type = "default",
  merge_form_name = "merged",
  filter_field = NULL,
  filter_choices = NULL,
  filter_list = NULL,
  filter_strict = TRUE,
  field_names = NULL,
  form_names = NULL,
  exclude_identifiers = TRUE,
  exclude_free_text = FALSE,
  date_handling = "none",
  labelled = TRUE,
  clean = TRUE,
  drop_blanks = FALSE,
  drop_missing_codes = FALSE,
  drop_others = NULL,
  include_metadata = TRUE,
  include_users = TRUE,
  include_records = TRUE,
  include_log = FALSE,
  annotate_from_log = TRUE,
  internal_use = FALSE
)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{summary_name}{Character. The name of the summary from which to generate
the summary. *If you provide \code{summary_name} all other parameters are
inherited according to what was set with \code{add_project_summary}.}

\item{transformation_type}{Character vector.
Default is "default".
Also have "none", "flat", "merge_non_repeating"
"default" first merges non-repeating and if there are repeating forms it
merges non-repeating variables to the right of repeating instruments
"flat" is one-record, one-row, even if there are repeating forms
"none" does not transform anything
"merge_non_repeating" still merges all non-repeating instruments but
does not merge them to repeating instruments}

\item{merge_form_name}{A character string representing the name of the merged
form. Default is "merged".}

\item{filter_field}{Character. The name of the field in the database to
filter on.}

\item{filter_choices}{Vector. The values of \code{filter_field} used to define the
summary.}

\item{filter_list}{Vector. The values of \code{filter_field} used to define the
summary.}

\item{filter_strict}{Logical. If \code{TRUE}, all forms will be filtered by
criteria. If \code{FALSE}, will convert original filter to id column and filter
all other forms by that record. Default is \code{TRUE}.}

\item{field_names}{Character vector. Names of specific fields to include in
the summary. Default is \code{NULL}, which includes all fields.}

\item{form_names}{Character vector. Names of forms to include in the summary.
Default is \code{NULL}, which includes all forms.}

\item{exclude_identifiers}{Logical. Whether to exlude identifiers in the data
in the summary. Default is \code{TRUE}.}

\item{exclude_free_text}{Logical for excluding free text. Default is \code{FALSE}.}

\item{date_handling}{character string. One of \code{none},\code{exclude_dates},
\code{random_shift_by_record}, \code{random_shift_by_project}, \code{zero_by_record}, or
\code{zero_by_project} random shift is +/- 90 unless changed with options}

\item{labelled}{Logical. If \code{TRUE}, the data will be converted to labelled.
Default is \code{TRUE}.}

\item{clean}{Logical. If \code{TRUE}, the data will be cleaned before summarizing.
Default is \code{TRUE}. If missing codes present AND number or date type, R will
convert to those to NA and would make that variable not upload compatible}

\item{drop_blanks}{Logical. If \code{TRUE}, records with blank fields will be
dropped. Default is \code{TRUE}.}

\item{drop_missing_codes}{Logical. If \code{TRUE}, will convert missing codes to NA.
Default is \code{FALSE}.}

\item{drop_others}{Character vector of other values that should be dropped.}

\item{include_metadata}{Logical. If \code{TRUE}, metadata will be included in the
summary. Default is \code{TRUE}.}

\item{include_users}{Logical. If \code{TRUE}, user-related information will be
included in the summary. Default is \code{TRUE}.}

\item{include_records}{Logical. If \code{TRUE}, a record summary will be
included in the generated summary. Default is \code{TRUE}.}

\item{include_log}{Logical. If \code{TRUE}, the log of changes will be included in
the summary. Default is \code{TRUE}.}

\item{annotate_from_log}{Logical. If \code{TRUE}, the metadata, users, and records
will be annotated using the log. Default is \code{TRUE}.}

\item{internal_use}{A logical flag (\code{TRUE} or \code{FALSE}). If \code{TRUE}, then will
return data_list meant for internal use. Defaults to \code{FALSE}.}
}
\value{
A list containing the generated summary based on the specified options. The
list includes filtered and cleaned data, metadata, and other summary details.
}
\description{
Generates a summary from a predefined summary of data within a REDCap project.
The summary can be customized based on various options, such as cleaning the
data, including metadata, and annotating metadata.
}
\details{
This function allows you to generate a summary of data from a specific summary
of records within the REDCap project. The function provides flexible options
for cleaning, annotating, and including metadata, as well as controlling
whether to include record summaries, user information, and logs.
}
\keyword{internal}
