% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_projects}
\alias{get_projects}
\title{Get your REDCap projects used by REDCapSync}
\usage{
get_projects()
}
\value{
data.frame of projects from the cache
}
\description{
Everytime a setup or update is performed REDCapSync stores the most basic
information
about that project to the cache so the user has a running log of everywhere
there project information is stored,
which can be used to find, move, edit, delete that data.
}
\seealso{
Other Project Cache Functions: 
\code{\link{cache_clear}()},
\code{\link{cache_path}()},
\code{\link{cache_remove_project}()}
}
\concept{Project Cache Functions}
\keyword{Cache}
\keyword{Functions}
\keyword{Project}
