% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_API.R
\name{get_redcap_report}
\alias{get_redcap_report}
\title{Get REDCap Report}
\usage{
get_redcap_report(project, report_id, silent = TRUE)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{report_id}{character or integer of REDCap report ID. This can be found
at the end of the URL of the report.}

\item{silent}{Logical (TRUE/FALSE). For messages.}
}
\value{
data.frame of REDCap report
}
\description{
Get REDCap Report
}
\keyword{internal}
