% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-project.R
\name{project}
\alias{project}
\title{Setup or Load REDCapSync Project}
\value{
A \code{project} R6 object.
}
\description{
Project class (REDCapSync)
Main class for managing REDCap data, metadata, and sync operations.
Users should construct objects using \code{\link[=setup_project]{setup_project()}}
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-project-new}{\code{project$new()}}
\item \href{#method-project-info}{\code{project$info()}}
\item \href{#method-project-add_summary}{\code{project$add_summary()}}
\item \href{#method-project-generate_summary}{\code{project$generate_summary()}}
\item \href{#method-project-add_field}{\code{project$add_field()}}
\item \href{#method-project-sync}{\code{project$sync()}}
\item \href{#method-project-save}{\code{project$save()}}
\item \href{#method-project-use}{\code{project$use()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project-new"></a>}}
\if{latex}{\out{\hypertarget{method-project-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this project  \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{project$new(short_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{short_name}}{Character project identifier.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project-info"></a>}}
\if{latex}{\out{\hypertarget{method-project-info}{}}}
\subsection{Method \code{info()}}{
Print project metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{project$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project-add_summary"></a>}}
\if{latex}{\out{\hypertarget{method-project-add_summary}{}}}
\subsection{Method \code{add_summary()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{project$add_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project-generate_summary"></a>}}
\if{latex}{\out{\hypertarget{method-project-generate_summary}{}}}
\subsection{Method \code{generate_summary()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{project$generate_summary(short_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{short_name}}{Character project identifier.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project-add_field"></a>}}
\if{latex}{\out{\hypertarget{method-project-add_field}{}}}
\subsection{Method \code{add_field()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{project$add_field()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project-sync"></a>}}
\if{latex}{\out{\hypertarget{method-project-sync}{}}}
\subsection{Method \code{sync()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{project$sync()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project-save"></a>}}
\if{latex}{\out{\hypertarget{method-project-save}{}}}
\subsection{Method \code{save()}}{
Add a new summary entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{project$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project-use"></a>}}
\if{latex}{\out{\hypertarget{method-project-use}{}}}
\subsection{Method \code{use()}}{
returns internal list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{project$use()}\if{html}{\out{</div>}}
}

}
}
