% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapDB-package.R
\docType{package}
\name{REDCapDB-package}
\alias{REDCapDB}
\alias{REDCapDB-package}
\title{REDCapDB: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thecodingdocs/REDCapDB}
  \item \url{https://thecodingdocs.github.io/REDCapDB/}
  \item Report bugs at \url{https://github.com/thecodingdocs/REDCapDB/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
