% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapDB-package.R
\docType{package}
\name{REDCapDB-package}
\alias{REDCapDB}
\alias{REDCapDB-package}
\title{REDCapDB: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
