% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapSync-package.R
\docType{package}
\name{REDCapSync-package}
\alias{REDCapSync}
\alias{REDCapSync-package}
\title{REDCapSync: Encapsulated REDCap Project Objects for Data Pipelines in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools to download, organize, and synchronize data from REDCap projects through the REDCap application programming interface (API). The package constructs a standardized project object that stores data, metadata, logs, and configuration details in a user-defined directory. It supports incremental updates by checking the REDCap log and retrieving only modified data. Full updates are only triggered by major metadata changes. User can define subsets, derived fields, and transformations, which are stored in the list object and passed down to output objects to be used for data pipelines and downstream applications, such as the R shiny app 'RosyREDCap'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thecodingdocs/REDCapSync}
  \item \url{https://thecodingdocs.github.io/REDCapSync/}
  \item Report bugs at \url{https://github.com/thecodingdocs/REDCapSync/issues}
}

}
\author{
\strong{Maintainer}: Brandon Rose \email{thecodingdocs@gmail.com} (\href{https://orcid.org/0009-0009-7813-1960}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Natalie Goulett [contributor]
  \item Chris Hilsinger-Pate [contributor]
  \item Nikola Susic [contributor]
  \item Mason Thornton [contributor]
  \item Sydney Stern (\href{https://orcid.org/0000-0002-1479-605X}{ORCID}) [contributor]
}

}
\keyword{internal}
