% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapSync-package.R
\docType{package}
\name{REDCapSync-package}
\alias{REDCapSync}
\alias{REDCapSync-package}
\title{REDCapSync: Synchronize Data Pipelines from the REDCap API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Uses the REDCap API to build a standardized R list object that can be stored in a user-defined local or cloud directory. Checks the REDCap log and only gets necessary data updates. Full updates are only triggered by major metadata changes. User can define subsets, derived fields, and transformations, which are stored in the list object and passed down to output objects to be used downstream for data pipelines and the R shiny app RosyREDCap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thecodingdocs/REDCapSync}
  \item \url{https://thecodingdocs.github.io/REDCapSync/}
  \item Report bugs at \url{https://github.com/thecodingdocs/REDCapSync/issues}
}

}
\author{
\strong{Maintainer}: Brandon Rose \email{thecodingdocs@gmail.com} (\href{https://orcid.org/0009-0009-7813-1960}{ORCID})

Other contributors:
\itemize{
  \item Natalie Goulett [contributor]
  \item Chris Hilsinger-Pate [contributor]
  \item Nikola Susic [contributor]
  \item Sydney Stern (\href{https://orcid.org/0000-0002-1479-605X}{ORCID}) [contributor]
}

}
\keyword{internal}
