% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapSync-package.R
\docType{package}
\name{REDCapSync-package}
\alias{REDCapSync}
\alias{REDCapSync-package}
\title{REDCapSync: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Uses the REDCap API to build a standardized R list object (DB), designed to be used downstream for data pipelines and the R shiny app RosyREDCap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thecodingdocs/REDCapSync}
  \item \url{https://thecodingdocs.github.io/REDCapSync/}
  \item Report bugs at \url{https://github.com/thecodingdocs/REDCapSync/issues}
}

}
\author{
\strong{Maintainer}: Brandon Rose \email{thecodingdocs@gmail.com} (\href{https://orcid.org/0009-0009-7813-1960}{ORCID})

Authors:
\itemize{
  \item Natalie Goulett
  \item Chris Hilsinger-Pate
  \item Nikola Susic
}

Other contributors:
\itemize{
  \item Sydney Stern (\href{https://orcid.org/0000-0002-1479-605X}{ORCID}) [contributor]
}

}
\keyword{internal}
