% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync}
\alias{sync}
\title{Synchronize your REDCaps}
\usage{
sync(project_names = NULL, hard_reset = FALSE, summarize = TRUE)
}
\arguments{
\item{project_names}{character vector of project short_names to check for
sync. Default is NULL which will check every project in cache.}

\item{hard_reset}{Will go get all projects from scratch if TRUE.}

\item{summarize}{Logical (TRUE/FALSE). If TRUE, summarizes data to directory.}
}
\description{
Synchronize your REDCaps
}
