% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_project.R
\name{sync}
\alias{sync}
\title{Synchronize REDCap Data}
\usage{
sync(
  project_names = NULL,
  summarize = TRUE,
  hard_check = FALSE,
  hard_reset = FALSE
)
}
\arguments{
\item{project_names}{character vector of project project_names previously
setup. If NULL, will get all from \code{get_projects()}}

\item{summarize}{Logical (TRUE/FALSE). If TRUE, summarizes data to directory.}

\item{hard_check}{Will check REDCap even if not due (see \code{sync_frequency}
parameter from \code{setup_project()})}

\item{hard_reset}{Logical that forces a fresh update if TRUE. Default is
\code{FALSE}.}
}
\value{
invisible return of last project
}
\description{
Syncs with REDCap via \code{project} object that user defined with
\link{setup_project}
}
\details{
Syncs all projects by default but can be used to hands-free sync one or
defined set projects. This is not intended to return project object. User
should use \code{load_project("project_name")}. However, by default will invisibly
return the last project in the set of project_names.
#' See \link{REDCapSync_project}
or see \href{../../REDCapSync/html/REDCapSync_project.html#method-sync}{\code{REDCapSync_project$sync()}}
or see \href{../../REDCapSync/reference/REDCapSync_project.html#method-REDCapSync_project-sync}{\code{REDCapSync_project$sync()}}
or see \href{../../REDCapSync/reference/REDCapSync_project.html#method-sync-}{\code{REDCapSync_project$sync()}}
}
\seealso{
\link{setup_project} for initializing the \code{project} object.
}
