% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_project.R
\name{sync_multiple}
\alias{sync_multiple}
\title{Synchronize REDCap Data}
\usage{
sync_multiple(
  short_names = NULL,
  summarize = TRUE,
  hard_check = FALSE,
  hard_reset = FALSE
)
}
\arguments{
\item{short_names}{character vector of project short_names previously setup.
If = NULL, will get all from \code{get_projects()}}

\item{summarize}{Logical (TRUE/FALSE). If TRUE, summarizes data to directory.}

\item{hard_check}{Will check REDCap even if not due (see \code{sync_frequency}
parameter from \code{setup_project()})}

\item{hard_reset}{Logical that forces a fresh update if TRUE. Default is
\code{FALSE}.}
}
\value{
invisible
}
\description{
Updates the REDCap database (\code{project} object) by fetching the latest data
from the REDCap server.
}
\details{
syncs multiple projects as defined. Will not load projects
}
\seealso{
\link{setup_project} for initializing the \code{project} object.
}
