% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_project.R
\name{sync_project}
\alias{sync_project}
\title{Synchronize REDCap Data}
\usage{
sync_project(
  project,
  summarize = TRUE,
  save_to_dir = TRUE,
  hard_check = FALSE,
  hard_reset = FALSE,
  silent = FALSE
)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data
and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{summarize}{Logical (TRUE/FALSE). If TRUE, summarizes data to directory.}

\item{save_to_dir}{Logical (TRUE/FALSE). If TRUE, saves the updated data to
the directory. Default is \code{TRUE}.}

\item{hard_check}{Will check REDCap even if not due (see \code{sync_frequency}
parameter from \code{setup_project()})}

\item{hard_reset}{Logical that forces a fresh update if TRUE. Default is \code{FALSE}.}

\item{silent}{Logical (TRUE/FALSE). For messages.}
}
\value{
Messages for confirmation.
}
\description{
Updates the REDCap database (\code{project} object) by fetching the latest data
from the REDCap server.
}
\details{
This function updates the REDCap database by fetching the latest data from
the REDCap server. It supports various options such as forcing a fresh
update, checking logs for a specified number of days, and retrieving files
from REDCap. The function can also handle metadata-only updates and batch
processing.
}
\seealso{
\link{setup_project} for initializing the \code{project} object.

Other db_functions: 
\code{\link{sync_all}()}
}
\concept{db_functions}
