% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_update.R
\name{sync_project}
\alias{sync_project}
\title{Syncronize REDCap Data}
\usage{
sync_project(
  project,
  set_token_if_fails = TRUE,
  reset = FALSE,
  day_of_log = 10,
  get_files = FALSE,
  original_file_names = FALSE,
  entire_log = FALSE,
  metadata_only = FALSE,
  ask_about_overwrites = TRUE,
  save_to_dir = TRUE,
  records = NULL,
  silent = FALSE
)
}
\arguments{
\item{project}{A validated \code{project} object containing REDCap project data and
settings. Generated using \link{load_project} or \link{setup_project}}

\item{set_token_if_fails}{Logical (TRUE/FALSE). If TRUE, prompts the user to
set the REDCap API token if the update fails. Default is \code{TRUE}.}

\item{reset}{Logical that forces a fresh update if TRUE. Default is \code{FALSE}.}

\item{day_of_log}{Integer. Number of days to be checked in the log. Default
is \code{10}.}

\item{get_files}{Logical (TRUE/FALSE). If TRUE, retrieves files from REDCap.
Default is \code{FALSE}.}

\item{original_file_names}{Logical (TRUE/FALSE). If TRUE, uses original file
names for retrieved files. Default is \code{FALSE}.}

\item{entire_log}{Logical (TRUE/FALSE). If TRUE, retrieves the entire log.
Default is \code{FALSE}.}

\item{metadata_only}{Logical (TRUE/FALSE). If TRUE, updates only the
metadata. Default is \code{FALSE}.}

\item{ask_about_overwrites}{Logical (TRUE/FALSE). If TRUE, prompts the user
before overwriting existing data. Default is \code{TRUE}.}

\item{save_to_dir}{Logical (TRUE/FALSE). If TRUE, saves the updated data to
the directory. Default is \code{TRUE}.}

\item{records}{optional records character vector}

\item{silent}{Logical (TRUE/FALSE). For messages.}
}
\value{
Messages for confirmation.
}
\description{
Updates the REDCap database (\code{project} object) by fetching the latest data from
the REDCap server.
}
\details{
This function updates the REDCap database by fetching the latest data from
the REDCap server. It supports various options such as forcing a fresh
update, checking logs for a specified number of days, and retrieving files
from REDCap. The function can also handle metadata-only updates and batch
processing.
}
\seealso{
\link{setup_project} for initializing the \code{project} object.

Other db_functions: 
\code{\link{rmarkdown_project}()},
\code{\link{summarize_project}()},
\code{\link{transform_project}()}
}
\concept{db_functions}
